+++ HobbyMatch.BL/DTOs/Venues/VenueDetailsDto.cs
@@ -13,3 +13,12 @@ public record VenueDetailsDto(
     int BusinessClientId,
     string ClientName
 );
+
+public static partial class VenueExtensions
+{
+    public static VenueDetailsDto ToDetailsDto(this Venue venue)
+    {
+        return new VenueDetailsDto(venue.Id, venue.Name, venue.Address, venue.MaxUsers, venue.Price, venue.Location,
+            venue.Description, venue.BusinessClientId, venue.BusinessClient.UserName!);
+    }
+}
